requirements.txt
pillow
pytesseract
fastapi
pytest
httpx
fastapi[testclient]
